Rename `configStore` to `configService`
